rename files with matrix_partition
