+++ cpp/include/cugraph/prims/transform_reduce_by_src_dst_key_e.cuh
@@ -17,7 +17,7 @@
 
 #include <cugraph/detail/graph_utils.cuh>
 #include <cugraph/graph_view.hpp>
-#include <cugraph/matrix_partition_device_view.cuh>
+#include <cugraph/edge_partition_device_view.cuh>
 #include <cugraph/prims/property_op_utils.cuh>
 #include <cugraph/utilities/dataframe_buffer.cuh>
 #include <cugraph/utilities/error.hpp>
