+++ cpp/include/cugraph/prims/transform_reduce_e.cuh
@@ -16,8 +16,8 @@
 #pragma once
 
 #include <cugraph/graph_view.hpp>
-#include <cugraph/matrix_partition_device_view.cuh>
-#include <cugraph/matrix_partition_view.hpp>
+#include <cugraph/edge_partition_device_view.cuh>
+#include <cugraph/edge_partition_view.hpp>
 #include <cugraph/prims/property_op_utils.cuh>
 #include <cugraph/utilities/dataframe_buffer.cuh>
 #include <cugraph/utilities/error.hpp>
