+++ cpp/include/cugraph/prims/update_frontier_v_push_if_out_nbr.cuh
@@ -16,7 +16,7 @@
 #pragma once
 
 #include <cugraph/graph_view.hpp>
-#include <cugraph/matrix_partition_device_view.cuh>
+#include <cugraph/edge_partition_device_view.cuh>
 #include <cugraph/partition_manager.hpp>
 #include <cugraph/prims/property_op_utils.cuh>
 #include <cugraph/prims/reduce_op.cuh>
