+++ cpp/tests/prims/mg_count_if_e.cu
@@ -27,7 +27,7 @@
 
 #include <cuco/detail/hash_functions.cuh>
 #include <cugraph/graph_view.hpp>
-#include <cugraph/matrix_partition_view.hpp>
+#include <cugraph/edge_partition_view.hpp>
 #include <cugraph/prims/count_if_e.cuh>
 #include <cugraph/prims/edge_partition_src_dst_property.cuh>
 #include <cugraph/prims/update_edge_partition_src_dst_property.cuh>
