+++ cpp/tests/prims/mg_update_frontier_v_push_if_out_nbr.cu
@@ -27,7 +27,7 @@
 
 #include <cuco/detail/hash_functions.cuh>
 #include <cugraph/graph_view.hpp>
-#include <cugraph/matrix_partition_view.hpp>
+#include <cugraph/edge_partition_view.hpp>
 #include <cugraph/prims/edge_partition_src_dst_property.cuh>
 #include <cugraph/prims/update_edge_partition_src_dst_property.cuh>
 #include <cugraph/prims/update_frontier_v_push_if_out_nbr.cuh>
